+++ src/components/molecule/editor/Editor.js
@@ -2,12 +2,25 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ReactQuill from 'react-quill';
 import './Editor.scss';
+import classNames from 'classnames';
 
-const Editor = ({ value, modules, onChange, toolbar, ...otherProps }) => {
+const Editor = ({
+  value,
+  modules,
+  fullHeight,
+  onChange,
+  toolbar,
+  ...otherProps
+}) => {
   const optionsToolbar = modules || {};
 
+  const classes = classNames({
+    'c-editor': true,
+    'c-editor--full-height': fullHeight
+  });
+
   return (
-    <div className="c-editor">
+    <div className={classes}>
       <ReactQuill
         modules={!toolbar ? { toolbar: null } : { toolbar: optionsToolbar }}
         theme="snow"
@@ -27,13 +40,15 @@ Editor.propTypes = {
     PropTypes.string,
     PropTypes.object,
     PropTypes.array
-  ])
+  ]),
+  fullHeight: PropTypes.bool
 };
 
 Editor.defaultProps = {
   value: '',
   toolbar: true,
-  modules: null
+  modules: null,
+  fullHeight: false
 };
 
 export default Editor;